chore(sql): refactored tests to remove even more duplicate code
