+++ core/src/test/java/io/questdb/mp/TestWorkerPool.java
@@ -1,0 +1,54 @@
+/*******************************************************************************
+ *     ___                  _   ____  ____
+ *    / _ \ _   _  ___  ___| |_|  _ \| __ )
+ *   | | | | | | |/ _ \/ __| __| | | |  _ \
+ *   | |_| | |_| |  __/\__ \ |_| |_| | |_) |
+ *    \__\_\\__,_|\___||___/\__|____/|____/
+ *
+ *  Copyright (c) 2014-2019 Appsicle
+ *  Copyright (c) 2019-2022 QuestDB
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ ******************************************************************************/
+
+package io.questdb.mp;
+
+import io.questdb.Metrics;
+import io.questdb.test.tools.TestUtils;
+
+public class TestWorkerPool extends WorkerPool {
+
+    public TestWorkerPool(int workerCount) {
+        this(workerCount, Metrics.disabled());
+    }
+
+    public TestWorkerPool(int workerCount, Metrics metrics) {
+        super(new WorkerPoolConfiguration() {
+            @Override
+            public int[] getWorkerAffinity() {
+                return TestUtils.getWorkerAffinity(workerCount);
+            }
+
+            @Override
+            public int getWorkerCount() {
+                return workerCount;
+            }
+
+            @Override
+            public boolean haltOnError() {
+                return false;
+            }
+        }, metrics);
+    }
+}
